config/requirements.txt
# CengBot - Çukurova Üniversitesi AI Chatbot Requirements

# Core AI Dependencies
torch>=2.0.0
transformers>=4.35.0
accelerate>=0.24.0
peft>=0.6.0
bitsandbytes>=0.41.0

# Telegram Bot
python-telegram-bot==20.7

# Web Framework & API
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6

# Database (SQLite - No PostgreSQL needed)
sqlalchemy>=2.0.41

# Message Queue
pika==1.3.2

# Language Detection
langdetect>=1.0.9

# Machine Learning & Similarity
scikit-learn>=1.3.0
numpy>=1.24.0
sentence-transformers>=2.2.0

# Data Augmentation
anthropic>=0.35.0

# Utilities
aiohttp==3.9.1
requests>=2.31.0
python-dotenv>=1.0.0

# System Requirements:
# - Python 3.8+
# - Node.js 16+ (for React frontend: cd admin_frontend && npm install)
# - RabbitMQ Server (sudo apt install rabbitmq-server)
# - CUDA compatible GPU (recommended for LLaMA 3.2 3B model)
# - Minimum 8GB RAM (16GB+ recommended for model inference)
#
# Installation:
# pip install -r requirements.txt
#
# Frontend:
# cd admin_frontend && npm install